+++ src/utils/triangle_numbers.rs
@@ -1,4 +1,4 @@
-pub struct TriangleNumbers(u64, u64);
+pub struct TriangleNumbers(usize, usize);
 
 impl TriangleNumbers {
     pub fn new() -> TriangleNumbers {
@@ -7,8 +7,8 @@ impl TriangleNumbers {
 }
 
 impl Iterator for TriangleNumbers {
-    type Item = u64;
-    fn next(&mut self) -> Option<u64> {
+    type Item = usize;
+    fn next(&mut self) -> Option<usize> {
         self.1 += 1;
         self.0 += self.1;
         Some(self.0)
